requirements.txt
streamlit
python-dotenv
pytz
langchain
langchain-community
langchain-huggingface
langchain-openai
unstructured
faiss-cpu
openai
PyPDF2
python-docx
tiktoken
pandas
plotly
langchain-text-splitters